ml-service/requirements.txt
# Core
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Kafka
kafka-python==2.0.2

# Database
psycopg2-binary==2.9.9

# Storage
boto3==1.34.0

# Image processing (для Mock и будущей модели)
pillow==10.1.0
trimesh==4.0.5
numpy==1.24.3

# Testing
pytest==7.4.3
pytest-mock==3.12.0
